pull all SQLite 3 code into DBManager to make it reusable with both open channel and group channel, DB is now a separate layer which makes code cleaner.
